refactor: split query schema responsibilities
