+++ wow-schema/src/test/kotlin/me/ahoo/wow/schema/query/JsonQuerySchemaSourceTest.kt
@@ -13,17 +13,11 @@
 
 package me.ahoo.wow.schema.query
 
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonSubTypes
-import com.fasterxml.jackson.annotation.JsonTypeInfo
-import com.fasterxml.jackson.annotation.JsonUnwrapped
-import io.swagger.v3.oas.annotations.media.Schema
 import me.ahoo.test.asserts.assert
 import me.ahoo.test.asserts.assertThrownBy
 import me.ahoo.wow.api.query.LogicalField
 import me.ahoo.wow.api.query.schema.QueryCardinality
 import me.ahoo.wow.api.query.schema.QueryModel
-import me.ahoo.wow.api.query.schema.QueryTemporal
 import me.ahoo.wow.api.query.schema.QueryValueType
 import me.ahoo.wow.api.query.schema.Temporal
 import me.ahoo.wow.modeling.MaterializedNamedAggregate
@@ -39,13 +33,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import reactor.core.publisher.Flux
 import reactor.core.scheduler.Schedulers
-import tools.jackson.core.JsonGenerator
-import tools.jackson.databind.SerializationContext
-import tools.jackson.databind.annotation.JsonSerialize
-import tools.jackson.databind.ser.std.StdSerializer
-import java.math.BigDecimal
-import java.time.Instant
-import java.time.LocalDate
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicInteger
@@ -502,279 +489,3 @@ class JsonQuerySchemaSourceTest {
         dynamicChildren = DeclarationValue.Set(false),
     )
 }
-
-@Schema(title = "State title", description = "State description")
-private data class StructuralState(
-    @field:Schema(
-        title = "Count title",
-        description = "Count description",
-        requiredMode = Schema.RequiredMode.REQUIRED,
-    )
-    val count: Int,
-    val ratio: BigDecimal,
-    val active: Boolean,
-    @field:Schema(requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    val optional: String? = null,
-    val status: StructuralStatus,
-    val address: StructuralAddress,
-    val items: List<StructuralItem>,
-    val tags: List<String>,
-)
-
-private enum class StructuralStatus { ACTIVE, INACTIVE }
-
-private data class StructuralAddress(val city: String)
-
-private data class StructuralItem(val quantity: Int)
-
-private data class JacksonState(
-    @field:JsonProperty("display_name")
-    val displayName: String,
-    @field:JsonProperty("display.name")
-    val dottedName: String,
-    @field:JsonProperty("display name")
-    val spacedName: String,
-    @field:JsonProperty("0")
-    val numericName: String,
-    @field:JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-    val secret: String,
-    @field:JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    val visible: String,
-    @get:JsonUnwrapped(prefix = "detail_", suffix = "_value")
-    val details: JacksonDetails,
-)
-
-private data class JacksonDetails(
-    @field:JsonProperty("nested")
-    val nested: String,
-)
-
-private data class CustomSerializerState(
-    val typeValue: TypeCustomSerializedValue,
-    @get:JsonSerialize(using = PropertyCustomSerializedValueSerializer::class)
-    val propertyValue: PropertyCustomSerializedValue,
-)
-
-@JsonSerialize(using = TypeCustomSerializedValueSerializer::class)
-private data class TypeCustomSerializedValue(val hidden: String)
-
-private class TypeCustomSerializedValueSerializer : StdSerializer<TypeCustomSerializedValue>(
-    TypeCustomSerializedValue::class.java,
-) {
-    override fun serialize(
-        value: TypeCustomSerializedValue,
-        generator: JsonGenerator,
-        provider: SerializationContext,
-    ) {
-        generator.writeString(value.hidden)
-    }
-}
-
-private data class PropertyCustomSerializedValue(val hidden: String)
-
-private class PropertyCustomSerializedValueSerializer : StdSerializer<PropertyCustomSerializedValue>(
-    PropertyCustomSerializedValue::class.java,
-) {
-    override fun serialize(
-        value: PropertyCustomSerializedValue,
-        generator: JsonGenerator,
-        provider: SerializationContext,
-    ) {
-        generator.writeString(value.hidden)
-    }
-}
-
-private data class CompositionState(
-    @field:Schema(allOf = [AllOfInherited::class])
-    val allOf: AllOfValue,
-    val anyOf: AnyOfValue,
-    @field:Schema(oneOf = [OneOfFirst::class, OneOfSecond::class])
-    val oneOf: OneOfValue,
-    val payment: PaymentValue,
-)
-
-private data class AllOfValue(val own: String)
-
-private data class AllOfInherited(val inherited: String)
-
-@Schema(anyOf = [AnyOfLeft::class, AnyOfRight::class])
-private interface AnyOfValue
-
-private data class AnyOfLeft(val left: String) : AnyOfValue
-
-private data class AnyOfRight(val right: String) : AnyOfValue
-
-private class OneOfValue
-
-private data class OneOfFirst(val first: String)
-
-private data class OneOfSecond(val second: String)
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "kind")
-@JsonSubTypes(
-    JsonSubTypes.Type(value = CardPayment::class, name = "card"),
-    JsonSubTypes.Type(value = BankPayment::class, name = "bank"),
-)
-private interface PaymentValue
-
-private data class CardPayment(val cardNumber: String) : PaymentValue
-
-private data class BankPayment(val account: String) : PaymentValue
-
-private data class RepeatedCompositionState(
-    @field:Schema(oneOf = [StringValueBranch::class, IntegerValueBranch::class])
-    val forward: RepeatedValue,
-    @field:Schema(oneOf = [IntegerValueBranch::class, StringValueBranch::class])
-    val reverse: RepeatedValue,
-)
-
-private class RepeatedValue
-
-private data class StringValueBranch(val value: String)
-
-private data class IntegerValueBranch(val value: Int)
-
-private data class PartiallyRequiredCompositionState(
-    @field:Schema(anyOf = [RequiredSharedValueBranch::class, OptionalSharedValueBranch::class])
-    val value: RepeatedValue,
-)
-
-private data class RequiredSharedValueBranch(val shared: String)
-
-private data class OptionalSharedValueBranch(
-    @field:Schema(requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    val shared: String,
-)
-
-private data class ConflictingAllOfValueTypesState(
-    @field:Schema(allOf = [StringValueBranch::class, IntegerValueBranch::class])
-    val value: RepeatedValue,
-)
-
-private data class NumericSubtypeAllOfValueTypesState(
-    @field:Schema(allOf = [IntegerValueBranch::class, DecimalValueBranch::class])
-    val forward: RepeatedValue,
-    @field:Schema(allOf = [DecimalValueBranch::class, IntegerValueBranch::class])
-    val reverse: RepeatedValue,
-)
-
-private data class DecimalValueBranch(val value: BigDecimal)
-
-private data class OpaqueAllOfValueTypesState(
-    @field:Schema(allOf = [OpaqueValueBranch::class, StringValueBranch::class])
-    val known: RepeatedValue,
-    @field:Schema(allOf = [OpaqueValueBranch::class, SecondOpaqueValueBranch::class])
-    val opaque: RepeatedValue,
-)
-
-private data class OpaqueValueBranch(val value: TypeCustomSerializedValue)
-
-private data class SecondOpaqueValueBranch(val value: TypeCustomSerializedValue)
-
-private data class ConflictingCompositionState(
-    @field:Schema(oneOf = [FirstTitledBranch::class, SecondTitledBranch::class])
-    val union: RepeatedValue,
-)
-
-private data class FirstTitledBranch(
-    @field:Schema(title = "First")
-    val value: String,
-)
-
-private data class SecondTitledBranch(
-    @field:Schema(title = "Second")
-    val value: String,
-)
-
-private data class ForwardMetadataState(
-    @field:Schema(oneOf = [FirstMetadataBranch::class, SecondMetadataBranch::class])
-    val value: RepeatedValue,
-)
-
-private data class ReverseMetadataState(
-    @field:Schema(oneOf = [SecondMetadataBranch::class, FirstMetadataBranch::class])
-    val value: RepeatedValue,
-)
-
-@Schema(title = "First title", description = "First description")
-private class FirstMetadataBranch
-
-@Schema(title = "Second title", description = "Second description")
-private class SecondMetadataBranch
-
-private data class ForwardEnumState(
-    @field:Schema(oneOf = [FirstChoice::class, SecondChoice::class])
-    val value: RepeatedValue,
-)
-
-private data class ReverseEnumState(
-    @field:Schema(oneOf = [SecondChoice::class, FirstChoice::class])
-    val value: RepeatedValue,
-)
-
-private enum class FirstChoice { FIRST, SHARED }
-
-private enum class SecondChoice { SECOND, SHARED }
-
-private data class EqualContainerMetadataState(
-    @field:Schema(oneOf = [SharedMetadataFirst::class, SharedMetadataSecond::class])
-    val metadata: RepeatedValue,
-    @field:Schema(oneOf = [LocalDate::class, Instant::class])
-    val temporal: RepeatedValue,
-)
-
-private data class MixedTemporalAlternativeState(
-    @field:Schema(anyOf = [LocalDate::class, String::class])
-    val anyOf: RepeatedValue,
-    @field:Schema(oneOf = [LocalDate::class, String::class])
-    val oneOf: RepeatedValue,
-)
-
-@Schema(title = "Shared title", description = "Shared description")
-private class SharedMetadataFirst
-
-@Schema(title = "Shared title", description = "Shared description")
-private class SharedMetadataSecond
-
-private data class RecursiveState(
-    val name: String,
-    val child: RecursiveState?,
-    val children: List<RecursiveState>,
-)
-
-private data class DeepLevelOne(val two: DeepLevelTwo)
-
-private data class DeepLevelTwo(val three: DeepLevelThree)
-
-private data class DeepLevelThree(val four: DeepLevelFour)
-
-private data class DeepLevelFour(val five: DeepLevelFive)
-
-private data class DeepLevelFive(val six: DeepLevelSix)
-
-private data class DeepLevelSix(val value: String)
-
-private data class DynamicState(
-    val attributes: Map<String, String>,
-    val attributeGroups: List<Map<String, String>>,
-    val closed: StructuralAddress,
-)
-
-private data class NativeTemporalState(
-    val date: LocalDate,
-    val instant: Instant,
-    val instants: List<Instant>,
-)
-
-private data class AnnotatedTemporalState(
-    @field:JsonProperty("created_at")
-    @field:QueryTemporal(TimeUnit.SECONDS)
-    val createdAt: Long,
-    @field:QueryTemporal
-    val timestamps: List<Long>,
-)
-
-private data class InvalidTemporalState(
-    @field:QueryTemporal
-    val createdAt: String,
-)
